agent/requirements.txt
# Зависимости для агента классификации токсичного контента

# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Transformers для DistilBERT
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0

# Валидация данных
pydantic==2.5.0

# Конфигурация
python-dotenv==1.0.0
